# project-root/products-service/src/main/resources/static/oas/product-service.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Product Service API",
    "version": "1.0.0",
    "description": "Contrato OAS del microservicio de productos"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local" }
  ],
  "paths": {
    "/products": {
      "get": {
        "summary": "Listar productos",
        "operationId": "getAllProducts",
        "responses": {
          "200": {
            "description": "Listado de productos",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear producto",
        "operationId": "createProduct",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ProductRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Product" } }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Obtener producto por id",
        "operationId": "getProductById",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "Producto", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
          "404": { "description": "No encontrado" }
        }
      },
      "put": {
        "summary": "Actualizar producto",
        "operationId": "updateProduct",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ProductRequest" } }
          }
        },
        "responses": {
          "200": { "description": "Producto actualizado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
          "404": { "description": "No encontrado" }
        }
      },
      "delete": {
        "summary": "Eliminar producto",
        "operationId": "deleteProduct",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "204": { "description": "Eliminado" },
          "404": { "description": "No encontrado" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number", "format": "double" },
          "stock": { "type": "integer", "format": "int32" }
        }
      },
      "ProductRequest": {
        "type": "object",
        "required": ["name", "price"],
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number", "format": "double" },
          "stock": { "type": "integer", "format": "int32" }
        }
      }
    }
  }
}